requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
gym>=0.21.0
stable-baselines3>=1.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0
tqdm>=4.64.0
rich>=12.0.0
python-dotenv 